created folders and splited into two components the overview components
